Changes  * Added sensors according to earlier reasoning (see below). Currently available for transport models only (isenthalpic, isentropic), in this case temperature sensors have to be modeled using a conditional structure.  * Minor modifications to avoid numeric Jacobians  * Added WaterSteamCycle1 example based on PowerFluid
Sensors in FluidSandbox:
Implementing sensors as separate components does not seem to be a good choice because, due to the largely varying properties available on the different connectors, simple sensor components do not work and more complex ones (e.g. flow temperature sensor) might degrade the robustness etc. of an approach. Currently, they are implemented as outputs on actual component base classes, that can be activated upon request (no unnecessary performance penalties).

In general, temperature sensor on storage elements can be implemented without conditional constructs (in contrast to transport elements where the mass flow direction / pressure drop sign has to be inquired to decide about the temperature). If the sensor reading is not a simple output it is thus advisable to use a temperature reading off a storage component (for closed loop control etc).